refactor: integrate ActionPointsExtractor service for action points extraction
- Replaced the native action points extraction logic with a dedicated ActionPointsExtractor service, improving code organization and maintainability.
- Enhanced error handling for invalid action points structure, ensuring robust response validation.
- Updated logging to provide clearer insights into the extraction process and potential errors.
- Adjusted the action points display to limit the number of urgent concerns shown, improving UI clarity.
